Refactor: Move TransactionListViewModel to commonMain and uncomment email sending logic
This commit moves the `TransactionListViewModel` from the Android-specific `app/src/main/java` directory to the common module `app/src/commonMain/kotlin/br/com/stonesdk/sdkdemo/ui/transactionList`.

Additionally, the previously commented-out email sending functionality in `EmailProviderWrapper` for both payment and cancellation emails has been uncommented, enabling these features. The Koin dependency injection module has been updated to reflect the new location of `TransactionListViewModel` and ensure correct dependency provision.
