+++ app/src/commonMain/kotlin/br/com/stonesdk/sdkdemo/di/KoinModule.kt
@@ -110,7 +110,8 @@ val commonModule: Module = module {
 
     viewModel {
         TransactionListViewModel(
-            transactionListProvider = get()
+            transactionProvider = get(),
+            emailProviderWrapper = get(),
         )
     }
