+++ app/src/commonMain/kotlin/br/com/stonesdk/sdkdemo/wrappers/EmailProviderWrapper.kt
@@ -1,5 +1,7 @@
 package br.com.stonesdk.sdkdemo.wrappers
 
+import br.com.stone.sdk.android.error.StoneStatus
+import co.stone.posmobile.sdk.callback.StoneResultCallback
 import co.stone.posmobile.sdk.payment.domain.model.response.PaymentData
 import co.stone.posmobile.sdk.payment.domain.model.response.TransactionStatus
 import co.stone.posmobile.sdk.sendEmail.domain.model.EmailConfig
@@ -40,48 +42,48 @@ class EmailProviderWrapper {
         config: EmailConfig,
         paymentData: PaymentData,
     ) {
-//        emailProvider.sendPaymentEmail(
-//            config = config,
-//            paymentData = paymentData,
-//            stoneResultCallback =
-//                object : StoneResultCallback<Unit> {
-//                    override fun onSuccess(result: Unit) {
-//                        trySend(EmailStatus.Success)
-//                    }
-//
-//                    override fun onError(
-//                        stoneStatus: StoneStatus?,
-//                        throwable: Throwable,
-//                    ) {
-//                        val error = stoneStatus?.message ?: throwable.message ?: "Unknown error"
-//                        trySend(EmailStatus.Error(error))
-//                    }
-//                },
-//        )
+        emailProvider.sendPaymentEmail(
+            config = config,
+            paymentData = paymentData,
+            stoneResultCallback =
+                object : StoneResultCallback<Unit> {
+                    override fun onSuccess(result: Unit) {
+                        trySend(EmailStatus.Success)
+                    }
+
+                    override fun onError(
+                        stoneStatus: StoneStatus?,
+                        throwable: Throwable,
+                    ) {
+                        val error = stoneStatus?.message ?: throwable.message ?: "Unknown error"
+                        trySend(EmailStatus.Error(error))
+                    }
+                },
+        )
     }
 
     private suspend fun ProducerScope<EmailStatus>.sendCancelEmail(
         config: EmailConfig,
         paymentData: PaymentData,
     ) {
-//        emailProvider.sendCancelEmail(
-//            config = config,
-//            initiatorTransactionKey = paymentData.initiatorTransactionKey,
-//            stoneResultCallback =
-//                object : StoneResultCallback<Unit> {
-//                    override fun onSuccess(result: Unit) {
-//                        trySend(EmailStatus.Success)
-//                    }
-//
-//                    override fun onError(
-//                        stoneStatus: StoneStatus?,
-//                        throwable: Throwable,
-//                    ) {
-//                        val error = stoneStatus?.message ?: throwable.message ?: "Unknown error"
-//                        trySend(EmailStatus.Error(error))
-//                    }
-//                },
-//        )
+        emailProvider.sendCancelEmail(
+            config = config,
+            initiatorTransactionKey = paymentData.initiatorTransactionKey,
+            stoneResultCallback =
+                object : StoneResultCallback<Unit> {
+                    override fun onSuccess(result: Unit) {
+                        trySend(EmailStatus.Success)
+                    }
+
+                    override fun onError(
+                        stoneStatus: StoneStatus?,
+                        throwable: Throwable,
+                    ) {
+                        val error = stoneStatus?.message ?: throwable.message ?: "Unknown error"
+                        trySend(EmailStatus.Error(error))
+                    }
+                },
+        )
     }
 
     private fun getEmailConfig(receiptType: EmailReceiptType): EmailConfig {
